requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.16.0
marshmallow==3.0.0rc4
Flask-Bcrypt==0.7.1
PyJWT==1.7.1
Flask-Migrate==2.3.1
Flask-Testing==0.7.1
Faker==1.0.2